Updating from new hot fixes
