+++ sdk/ml/azure-ai-ml/azure/ai/ml/_artifacts/_blob_storage_helper.py
@@ -6,26 +6,35 @@ import uuid
 import logging
 import time
 import os
-from typing import Dict, List, TYPE_CHECKING
+import warnings
+from contextlib import suppress
+from typing import Optional, Dict, Any, List, TYPE_CHECKING
 from pathlib import PurePosixPath, Path
+from multiprocessing import cpu_count
 from colorama import Fore
+from concurrent.futures import ThreadPoolExecutor, as_completed
+from tqdm import tqdm, TqdmWarning
+from platform import system
 import sys
 
+from azure.ai.ml._utils._exception_utils import EmptyDirectoryError
 from azure.storage.blob import BlobServiceClient, ContainerClient
 from azure.core.exceptions import ResourceNotFoundError
 from azure.ai.ml._utils._asset_utils import (
     generate_asset_id,
-    upload_directory,
-    upload_file,
+    traverse_directory,
     AssetNotChangedError,
     _build_metadata_dict,
     IgnoreFile,
+    FileUploadProgressBar,
     get_directory_size,
 )
 from azure.ai.ml._artifacts._constants import (
     UPLOAD_CONFIRMATION,
     ARTIFACT_ORIGIN,
     LEGACY_ARTIFACT_DIRECTORY,
+    EMPTY_DIRECTORY_ERROR,
+    PROCESSES_PER_CORE,
     MAX_CONCURRENCY,
     FILE_SIZE_WARNING,
     BLOB_DATASTORE_IS_HDI_FOLDER_KEY,
@@ -94,18 +103,11 @@ class BlobStorageClient:
 
             # start upload
             if os.path.isdir(source):
-                upload_directory(
-                    storage_client=self,
-                    source=source,
-                    dest=asset_id,
-                    msg=msg,
-                    show_progress=show_progress,
-                    ignore_file=ignore_file,
-                )
+                self.upload_dir(source, asset_id, msg, show_progress, ignore_file=ignore_file)
             else:
                 self.indicator_file = dest
                 self.check_blob_exists()
-                upload_file(storage_client=self, source=source, dest=dest, msg=msg, show_progress=show_progress)
+                self.upload_file(source, dest, msg, show_progress)
             print(Fore.RESET + "\n", file=sys.stderr)
 
             # upload must be completed before we try to generate confirmation file
@@ -121,6 +123,98 @@ class BlobStorageClient:
         artifact_info = {"remote path": dest, "name": name, "version": version, "indicator file": self.indicator_file}
 
         return artifact_info
+
+    def upload_file(
+        self,
+        source: str,
+        dest: str,
+        msg: Optional[str] = None,
+        show_progress: Optional[bool] = None,
+        in_directory: bool = False,
+        callback: Any = None,
+    ) -> None:
+        """
+        Upload a single file to a path inside the container
+        """
+        validate_content = os.stat(source).st_size > 0  # don't do checksum for empty files
+
+        with open(source, "rb") as data:
+            if show_progress and not in_directory:
+                file_size, _ = get_directory_size(source)
+                file_size_in_mb = file_size / 10**6
+                if file_size_in_mb < 1:
+                    msg += Fore.GREEN + " (< 1 MB)"
+                else:
+                    msg += Fore.GREEN + f" ({round(file_size_in_mb, 2)} MBs)"
+                cntx_manager = FileUploadProgressBar(msg=msg)
+            else:
+                cntx_manager = suppress()
+
+            with cntx_manager as c:
+                callback = c.update_to if (show_progress and not in_directory) else None
+                self.container_client.upload_blob(
+                    name=dest,
+                    data=data,
+                    validate_content=validate_content,
+                    overwrite=self.overwrite,
+                    raw_response_hook=callback,
+                    max_concurrency=MAX_CONCURRENCY,
+                )
+
+        self.uploaded_file_count += 1
+
+    def upload_dir(self, source: str, dest: str, msg: str, show_progress: bool, ignore_file: IgnoreFile) -> None:
+        """
+        Upload a directory to a path inside the container
+
+        Azure Blob doesn't allow metadata setting at the directory level, so the first
+        file in the directory is designated as the file where the confirmation metadata
+        will be added at the end of the upload.
+        """
+        source_path = Path(source).resolve()
+        prefix = "" if dest == "" else dest + "/"
+        prefix += os.path.basename(source_path) + "/"
+
+        # get all paths in directory and each file's size
+        upload_paths = []
+        size_dict = {}
+        total_size = 0
+        for root, _, files in os.walk(source_path):
+            upload_paths += list(traverse_directory(root, files, source_path, prefix, ignore_file=ignore_file))
+
+        for path, _ in upload_paths:
+            path_size = os.path.getsize(path)
+            size_dict[path] = path_size
+            total_size += path_size
+
+        upload_paths = sorted(upload_paths)
+        if len(upload_paths) == 0:
+            raise EmptyDirectoryError(
+                message=EMPTY_DIRECTORY_ERROR.format(source),
+                no_personal_data_message=msg.format("[source]"),
+                target=ErrorTarget.ARTIFACT,
+                error_category=ErrorCategory.USER_ERROR,
+            )
+
+        self.indicator_file = upload_paths[0][1]
+        self.check_blob_exists()
+        self.total_file_count = len(upload_paths)
+
+        # submit paths to workers for upload
+        num_cores = int(cpu_count()) * PROCESSES_PER_CORE
+        with ThreadPoolExecutor(max_workers=num_cores) as ex:
+            futures_dict = {
+                ex.submit(self.upload_file, src, dest, in_directory=True, show_progress=show_progress): (src, dest)
+                for (src, dest) in upload_paths
+            }
+            if show_progress:
+                warnings.simplefilter("ignore", category=TqdmWarning)
+                msg += f" ({round(total_size/10**6, 2)} MBs)"
+                ascii = system() == "Windows"  # Default unicode progress bar doesn't display well on Windows
+                with tqdm(total=total_size, desc=msg, ascii=ascii) as pbar:
+                    for future in as_completed(futures_dict):
+                        file_path_name = futures_dict[future][0]
+                        pbar.update(size_dict.get(file_path_name) or 0)
 
     def check_blob_exists(self) -> None:
         """
